# Make TPC-H Generation faster, increase default SF to 1