--- a/src/benchmarklib/abstract_table_generator.cpp
+++ b/src/benchmarklib/abstract_table_generator.cpp
@@ -59,12 +59,15 @@ void AbstractTableGenerator::generate_and_store() {
   /**
    * Add the Tables to the StorageManager
    */
-  _benchmark_config->out << "- Adding Tables to StorageManager" << std::endl;
+  _benchmark_config->out << "- Adding Tables to StorageManager (including generation of statistics) " << std::flush;
   auto& storage_manager = StorageManager::get();
   for (auto& [table_name, table_info] : table_info_by_name) {
     if (storage_manager.has_table(table_name)) storage_manager.drop_table(table_name);
     storage_manager.add_table(table_name, table_info.table);
   }
+
+  _benchmark_config->out << "(" << format_duration(std::chrono::duration_cast<std::chrono::nanoseconds>(timer.lap()))
+                         << ")" << std::endl;
 }
 
 }  // namespace opossum