--- a/src/benchmarklib/benchmark_runner.cpp
+++ b/src/benchmarklib/benchmark_runner.cpp
@@ -57,7 +57,6 @@ BenchmarkRunner::~BenchmarkRunner() {
 }
 
 void BenchmarkRunner::run() {
-  _config.out << "- Loading/Generating tables" << std::endl;
   _table_generator->generate_and_store();
 
   // Run the preparation queries