--- a/src/benchmarklib/benchmark_table_encoder.cpp
+++ b/src/benchmarklib/benchmark_table_encoder.cpp
@@ -132,19 +132,22 @@ bool BenchmarkTableEncoder::encode(const std::string& table_name, const std::sha
   auto encoding_performed = false;
   const auto column_data_types = table->column_data_types();
 
+  // Encode chunks in parallel, using `hardware_concurrency + 1` worker
   // Not using JobTasks here because we want parallelism even if the scheduler is disabled.
   auto next_chunk = std::atomic_uint{0};
   auto threads = std::vector<std::thread>{};
 
   for (auto thread_id = 0u; thread_id < std::thread::hardware_concurrency() + 1; ++thread_id) {
     threads.emplace_back([&] {
-      auto my_chunk = next_chunk++;
-      if (next_chunk >= table->chunk_count()) return;
+      while (true) {
+        auto my_chunk = next_chunk++;
+        if (next_chunk >= table->chunk_count()) return;
 
-      const auto& chunk = table->get_chunk(ChunkID{my_chunk});
-      if (!is_chunk_encoding_spec_satisfied(chunk_encoding_spec, get_chunk_encoding_spec(*chunk))) {
-        ChunkEncoder::encode_chunk(chunk, column_data_types, chunk_encoding_spec);
-        encoding_performed = true;
+        const auto& chunk = table->get_chunk(ChunkID{my_chunk});
+        if (!is_chunk_encoding_spec_satisfied(chunk_encoding_spec, get_chunk_encoding_spec(*chunk))) {
+          ChunkEncoder::encode_chunk(chunk, column_data_types, chunk_encoding_spec);
+          encoding_performed = true;
+        }
       }
     });
   }