--- a/src/benchmarklib/benchmark_table_encoder.cpp
+++ b/src/benchmarklib/benchmark_table_encoder.cpp
@@ -1,5 +1,6 @@
 #include "benchmark_table_encoder.hpp"
 
+#include <atomic>
 #include <thread>
 
 #include "constant_mappings.hpp"
@@ -129,7 +130,7 @@ bool BenchmarkTableEncoder::encode(const std::string& table_name, const std::sha
   /**
    * 2. Actually encode chunks
    */
-  auto encoding_performed = false;
+  auto encoding_performed = std::atomic<bool>{false};
   const auto column_data_types = table->column_data_types();
 
   // Encode chunks in parallel, using `hardware_concurrency + 1` worker