--- a/src/lib/statistics/generate_column_statistics.cpp
+++ b/src/lib/statistics/generate_column_statistics.cpp
@@ -1,5 +1,7 @@
 #include "generate_column_statistics.hpp"
 
+#include <boost/container/pmr/monotonic_buffer_resource.hpp>
+
 #include "storage/segment_iterate.hpp"
 
 namespace opossum {
@@ -11,15 +13,20 @@ namespace opossum {
 template <>
 std::shared_ptr<BaseColumnStatistics> generate_column_statistics<std::string>(const Table& table,
                                                                               const ColumnID column_id) {
-  std::unordered_set<std::string> distinct_set;
   // It would be nice to use string_view here, but the iterables hold copies of the values, not references themselves.
   // SegmentPosition would have to be changed to `T& _value` and this brings a whole bunch of problems in iterators
   // that create stack copies of the accessed values (e.g., for ReferenceSegments)
 
+  auto temp_buffer = boost::container::pmr::monotonic_buffer_resource(table.row_count() * 10);
+  auto distinct_set =
+      std::unordered_set<std::string, std::hash<std::string>, std::equal_to<std::string>,
+                         PolymorphicAllocator<std::string>>(PolymorphicAllocator<std::string>{&temp_buffer});
+  distinct_set.reserve(table.row_count());
+
   auto null_value_count = size_t{0};
 
-  auto min = std::string{};
-  auto max = std::string{};
+  auto min = std::string_view{};
+  auto max = std::string_view{};
 
   for (ChunkID chunk_id{0}; chunk_id < table.chunk_count(); ++chunk_id) {
     const auto base_segment = table.get_chunk(chunk_id)->get_segment(column_id);
@@ -28,14 +35,22 @@ std::shared_ptr<BaseColumnStatistics> generate_column_statistics<std::string>(co
       if (position.is_null()) {
         ++null_value_count;
       } else {
-        if (distinct_set.empty()) {
-          min = position.value();
-          max = position.value();
-        } else {
-          min = std::min(min, position.value());
-          max = std::max(max, position.value());
+        // One would expect distinct_set.emplace() to be the same as the code below. However, "The element may be
+        // constructed even if there already is an element with the key in the container, in which case the newly
+        // constructed element will be destroyed immediately."
+        // This is the case here, where simply using emplace takes ~50% longer.
+        auto it = distinct_set.find(position.value());
+        if (it == distinct_set.end()) {
+          it = distinct_set.emplace_hint(it, std::move(position.value()));
         }
-        distinct_set.insert(position.value());
+
+        if (distinct_set.size() == 1) {
+          min = *it;
+          max = *it;
+        } else {
+          if (*it < min) min = *it;
+          if (*it > max) max = *it;
+        }
       }
     });
   }
@@ -44,7 +59,8 @@ std::shared_ptr<BaseColumnStatistics> generate_column_statistics<std::string>(co
       table.row_count() > 0 ? static_cast<float>(null_value_count) / static_cast<float>(table.row_count()) : 0.0f;
   const auto distinct_count = static_cast<float>(distinct_set.size());
 
-  return std::make_shared<ColumnStatistics<std::string>>(null_value_ratio, distinct_count, min, max);
+  return std::make_shared<ColumnStatistics<std::string>>(null_value_ratio, distinct_count, std::string{min},
+                                                         std::string{max});
 }
 
 }  // namespace opossum