--- a/src/lib/statistics/generate_column_statistics.cpp
+++ b/src/lib/statistics/generate_column_statistics.cpp
@@ -19,7 +19,7 @@ std::shared_ptr<BaseColumnStatistics> generate_column_statistics<std::string>(co
 
   auto temp_buffer = boost::container::pmr::monotonic_buffer_resource(table.row_count() * 10);
   auto distinct_set =
-      std::unordered_set<std::string, std::hash<std::string>, std::equal_to<std::string>,
+      std::unordered_set<std::string, std::hash<std::string>, std::equal_to<>,
                          PolymorphicAllocator<std::string>>(PolymorphicAllocator<std::string>{&temp_buffer});
   distinct_set.reserve(table.row_count());