--- a/src/lib/statistics/generate_column_statistics.hpp
+++ b/src/lib/statistics/generate_column_statistics.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <boost/container/pmr/monotonic_buffer_resource.hpp>
+#include <boost/container/scoped_allocator.hpp>
 #include <unordered_set>
 
 #include "base_column_statistics.hpp"
@@ -15,7 +17,13 @@ namespace opossum {
  */
 template <typename ColumnDataType>
 std::shared_ptr<BaseColumnStatistics> generate_column_statistics(const Table& table, const ColumnID column_id) {
-  std::unordered_set<ColumnDataType> distinct_set;
+  // distinct_set is thrown away at the end of this method, so we don't want proper heap allocations for the strings
+  // stored within. The initial size of the buffer is a completely random guess, but better than zero.
+
+  auto temp_buffer = boost::container::pmr::monotonic_buffer_resource(table.row_count() * 10);
+  auto distinct_set =
+      std::unordered_set<ColumnDataType, std::hash<ColumnDataType>, std::equal_to<ColumnDataType>,
+                         PolymorphicAllocator<ColumnDataType>>(PolymorphicAllocator<ColumnDataType>{&temp_buffer});
 
   auto null_value_count = size_t{0};