--- a/src/lib/storage/value_segment.cpp
+++ b/src/lib/storage/value_segment.cpp
@@ -42,11 +42,22 @@ ValueSegment<T>::ValueSegment(const std::vector<T>& values, const PolymorphicAll
     : BaseValueSegment(data_type_from_type<T>()), _values(values, alloc) {}
 
 template <typename T>
+ValueSegment<T>::ValueSegment(std::vector<T>&& values, const PolymorphicAllocator<T>& alloc)
+    : BaseValueSegment(data_type_from_type<T>()), _values(std::move(values), alloc) {}
+
+template <typename T>
 ValueSegment<T>::ValueSegment(const std::vector<T>& values, std::vector<bool>& null_values,
                               const PolymorphicAllocator<T>& alloc)
     : BaseValueSegment(data_type_from_type<T>()),
       _values(values, alloc),
       _null_values(pmr_concurrent_vector<bool>(null_values, alloc)) {}
+
+template <typename T>
+ValueSegment<T>::ValueSegment(std::vector<T>&& values, std::vector<bool>&& null_values,
+                              const PolymorphicAllocator<T>& alloc)
+    : BaseValueSegment(data_type_from_type<T>()),
+      _values(std::move(values), alloc),
+      _null_values(pmr_concurrent_vector<bool>(std::move(null_values), alloc)) {}
 
 template <typename T>
 const AllTypeVariant ValueSegment<T>::operator[](const ChunkOffset chunk_offset) const {