--- a/src/lib/storage/value_segment.hpp
+++ b/src/lib/storage/value_segment.hpp
@@ -22,7 +22,10 @@ class ValueSegment : public BaseValueSegment {
   explicit ValueSegment(pmr_concurrent_vector<T>&& values, pmr_concurrent_vector<bool>&& null_values,
                         const PolymorphicAllocator<T>& alloc = {});
   explicit ValueSegment(const std::vector<T>& values, const PolymorphicAllocator<T>& alloc = {});
+  explicit ValueSegment(std::vector<T>&& values, const PolymorphicAllocator<T>& alloc = {});
   explicit ValueSegment(const std::vector<T>& values, std::vector<bool>& null_values,
+                        const PolymorphicAllocator<T>& alloc = {});
+  explicit ValueSegment(std::vector<T>&& values, std::vector<bool>&& null_values,
                         const PolymorphicAllocator<T>& alloc = {});
 
   // Return the value at a certain position. If you want to write efficient operators, back off!