--- a/src/lib/types.hpp
+++ b/src/lib/types.hpp
@@ -79,8 +79,7 @@ class pmr_concurrent_vector : public tbb::concurrent_vector<T> {
   pmr_concurrent_vector(const std::vector<T>& values, PolymorphicAllocator<T> alloc = {})  // NOLINT
       : tbb::concurrent_vector<T>(values.begin(), values.end()), _alloc(alloc) {}
   pmr_concurrent_vector(std::vector<T>&& values, PolymorphicAllocator<T> alloc = {})  // NOLINT
-      : tbb::concurrent_vector<T>(std::make_move_iterator(values.begin()),
-                                  std::make_move_iterator(values.end())),
+      : tbb::concurrent_vector<T>(std::make_move_iterator(values.begin()), std::make_move_iterator(values.end())),
         _alloc(alloc) {}
 
   template <class I>