Before (nemea, SF 1):
- Loading/Generating tables (15 s 808 ms)
- Encoding tables (24 s 226 ms)
- Adding Tables to StorageManager and generating statistics (~18 s)

Now:
- Loading/Generating tables (9 s 484 ms)
- Encoding tables (2 s 977 ms)
- Adding Tables to StorageManager and generating statistics (8 s 438 ms)