@lanice 2019-01-08
Okay, this is kinda weird: tpch_table_generator.cpp, l. 252:
```
auto customer = call_dbgen_mk<customer_t>(row_idx + 1, mk_cust, TpchTable::Customer);
```
throws a segfault sometimes. Like every other or every third time. Both in Release and Debug mode, but not with gdb. On pella, no other configs, just `./build/hyriseBenchmarkTPCH`.

This does not happen on master. Any idea where that might come from?