@mrks 2019-01-11
Can't reproduce. Let's see what asan/memcheck think about it.