@mrks 2019-01-11
Gna. Null-terminated strings are stupid.